+++ CHANGELOG.md
@@ -6,6 +6,18 @@ Changes are grouped under the labels: `Added`, `Changed`, `Deprecated`, `Fixed`,
 `Removed` and `Security`.
 
 ---
+## unpublished
+
+### Added
+- FileUpload component
+### Changed
+- Updates to support DS v4:
+  - Remove the 'iconPath' prop from DatePicker, which is no longer required in DS v4
+  - Remove a lot of @ts-expect-error now the Design System supplies TS types
+  - Remove call to DS Aspect Box script, which is being removed from the DS (it was required for IE11 support, which is being dropped)
+  - SiteSearch has improved type safety
+- Allow hintText and errorMessage props to accept either string or React.ReactNode in: Checkbox, DatePicker, Question, Radio, Select, Textarea, TetInput
+
 ## [0.13.0] - 2025-12-19
 
 ### Added